+++ js/data.js
@@ -1,5 +1,5 @@
 import {getRandomInteger, getProgressNumber, getRandomFromArr} from './util.js';
-import {getMesseges} from './get-messeges.js';
+
 const arrDescriptions = [
   'Утренний туман', 'Закат на пляже', 'Цветущее поле', 'Горный пейзаж', 'Прогулка в парке', 'Девушка на лужайке',
   'Лесной ручей', 'Красивый город', 'Пара на закате', 'Собака на пляже', 'Рассвет на озере', 'Дети на качелях',
@@ -13,14 +13,55 @@ const arrNames = [
   'Рикардо', 'Ромина', 'Сабрина', 'Сильвия', 'Фабио', 'Франческа'
 ];
 
+const arrMessages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
 
 const getIdPhoto = getProgressNumber(); // Функция возвращения случайного ID фото
 const getUrlPhoto = getProgressNumber(); // Функция возвращения случайного N URL
-
+const getIdComment = getProgressNumber(); // Функция возвращения случайного ID коммента
 
 /**
  *
- * @returns инфо-блок для 1 фотографииq
+ * @returns строку комментария
+ * @description создаём из 1 или 2 случайных фраз из массива
+ */
+const getMesseges = function() {
+  let strokeMesseges = '';
+  const countMessages = getRandomInteger(1, 2);
+  for(let j = 0; j < countMessages; j++) {
+    strokeMesseges += `${getRandomFromArr(arrMessages) } `;
+  }
+  return strokeMesseges.slice(0, -1);
+};
+
+/**
+ *
+ * @returns массив комментариев
+ */
+const getArrOfComents = function() {
+  const arrComments = [];
+  const getListComments = getRandomInteger(0, 30);
+
+  for(let i = 0; i < getListComments; i++) {
+    arrComments.push({
+      id: getIdComment(),
+      avatar: `img/avatar-${ getRandomInteger(1, 6) }.svg`,
+      message: getMesseges(),
+      name: getRandomFromArr(arrNames)
+    });
+  }
+  return arrComments;
+};
+
+/**
+ *
+ * @returns инфо-блок для 1 фотографии
  */
 const createInfoFoto = function() {
   return {
@@ -31,5 +72,6 @@ const createInfoFoto = function() {
     comments: getArrOfComents()
   };
 };
+
 
 export {createInfoFoto};
